+++ ZLPhotoLib/Classes/ZLCamera/ZLCameraViewController.m
@@ -442,7 +442,7 @@ static CGFloat BOTTOM_HEIGHT = 60;
 {
     // 判断图片的限制个数
     if (self.maxCount > 0 && self.images.count >= self.maxCount) {
-        UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:@"提示" message:[NSString stringWithFormat:@"拍照的个数不能超过%ld",self.maxCount]delegate:nil cancelButtonTitle:@"好的" otherButtonTitles:nil, nil];
+        UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:@"提示" message:[NSString stringWithFormat:@"拍照的个数不能超过%ld",(long)self.maxCount]delegate:nil cancelButtonTitle:@"好的" otherButtonTitles:nil, nil];
         [alertView show];
         return ;
     }
