+++ ZLPhotoLib/Classes/ZLPhotoBrowser/ViewControllers/ZLPhotoPickerBrowserViewController.m
@@ -532,7 +532,7 @@ static NSString *_cellIdentifier = @"collectionViewCell";
 
 
 -(void)setPageLabelPage:(NSInteger)page{
-    self.pageLabel.text = [NSString stringWithFormat:@"%ld / %ld",page + 1, self.photos.count];
+    self.pageLabel.text = [NSString stringWithFormat:@"%d / %ld",page + 1, (unsigned long)self.photos.count];
     if (self.isPush) {
         self.title = self.pageLabel.text;
     }
