+++ ZLPhotoLib/Classes/ZLPhotoPicker/Views/ZLPhotoPickerFooterCollectionReusableView.m
@@ -32,7 +32,7 @@
     _count = count;
     
     if (count > 0) {
-        self.footerLabel.text = [NSString stringWithFormat:@"有 %ld 张图片", (NSInteger)count];
+        self.footerLabel.text = [NSString stringWithFormat:@"有 %ld 张图片", (long)count];
     }
 }
